General cleanup of ZBEE ZCL dissector.  Removed unnecessary casts and wrapping functions.
Tried to remove zcl_cluster_id global, but it's used in a function that is called by other dissectors that I'm not sure if they are expecting the zcl_cluster_id to be "saved" (or if -1 could be passed in for cluster id from those dissectors).  Removed as many uses of it as possible if someone knows the dissectors better to make that call.

svn path=/trunk/; revision=54186