Since the documentation states that, "Text2pcap understands a hexdump of the form generated by od -Ax -tx1 -v.", supply an example that actually matches that format.
In addition, change the hex bytes so that they represent a valid, well-formed packet, rather than a malformed RSVP packet.  This should allow a user to more clearly verify that the text2pcap converstion worked as expected in the event he/she desires to test text2pcap using the example provided.


svn path=/trunk/; revision=54454