Filetap library, version 0.001
This is intended as the "tap library" for fileshark.  Right now its basically just a very stripped down (and renamed) copy of the wiretap library.  The goal is to remove "capture" and "wire" specific functionality out of the file handling to make it easier/simpler to support non-capture type files in the epan architecture.

svn path=/trunk/; revision=54531