In a source file that defines external functions, the header file that declares the functions must be included, in order to make sure the declarations match the function signature.  Make it so.
They *didn't* match; fix that.

Do *not* use 16-bit tvbuff offsets; there is no advantage to that (and,
in fact, 32-bit arithmetic may be easier on some platforms), and it has
the disadvantage that you're more likely to have the offset overflow.

dissect_cbs_serial_number() returns the new offset, not the length of
the item it dissected; don't add its return value to the offset, set the
offset to its return value.

Make dissect_cbs_message_identifier() return the new offset as well, for
consistency, and don't add *its* return value to the offset.

svn path=/trunk/; revision=54714