f_k() is used outside packet-gsm_a_rr.c, so it can't be static.
In a source file that defines external functions, the header file that
declares the functions must be included, in order to make sure the
declarations match the function signature.  Make it so - which means
creating the header file in the first place.

That means f_k() doesn't need to, and shouldn't be, declared in
packet-gsm_rlcmac.h.  Include packet-gsm_a_rr.h in packet-gsm_rlcmac.c.

svn path=/trunk/; revision=54724