To quote section B.5 "Keys for password integrity mode" of PKCS #12 v1.1: Personal Information Exchange Syntax:
	http://www.emc.com/emc-plus/rsa-labs/pkcs/files/h11301-wp-pkcs-12v1-1-personal-information-exchange-syntax.pdf

"When password integrity mode is used to protect a PFX PDU, a password
and salt are used to derive a MAC key.  As with password privacy mode,
the password is a Unicode string, and the salt is a byte string."

So, not having found any other references to salts as text strings, copy
it with tvb_memdup(), not tvb_get_string().

svn path=/trunk/; revision=54893