Fix display of right arrow in CSV export
Changeset 1d8a895fa4d7c16d213bda50b1177eccbb54326e introduced the use of UTF-8 righ arrow to indicate the direction in TCP dissector.
While it displays nicely in Wireshark GUI or in a text export of packets, an export to CSV results in an escaped string.
This patch is a naive attempt to display the right arrow in a more friendly way when exporting to CSV.
Any smarter fix is welcome.

Change-Id: Ife787268696fa69dafc24a5cf9706af4c4832831