Fix Bug 6489 to stop compiler warning on gcrypt deprecated delcarations
The libgcrypt library has several deprecated functions exported, that we don't use.
Unfortunately the GCC compiler warns about these deprecated functions regardless,
which is quite annoying.  This commit makes clang/gcc ignore the deprecated attribute,
for gcrypt.h only.  The danger with this is if gcrypt ever deprecates a function
we actually *use*, then we won't see the warning.  So I'm not sure if it's a good or
bad idea to do this change, but it's really annoying to see the warnings and makes
finding real warnings difficult.

Change-Id: I03e80a6e7e4833ce0f709088c9ab4af98193db3d
Reviewed-on: https://code.wireshark.org/review/128
Reviewed-by: Martin Kaiser <wireshark@kaiser.cx>
Tested-by: Martin Kaiser <wireshark@kaiser.cx>