Fix Bug 9728 'Lua: ProtoField.bool() VALUESTRING argument is not optional but was supposed to be'
Similar to bug 9725 and ProtoField.new(), the way the VALUESTRING argument is being checked
in the code for ProtoField.bool() ends up making it non-optional.  This patch fixes that,
along with some minor API documentation fixes (text).

This is a backport for change-id Iadb9a8ace9c5514fc623d882301fe16b637fe4ce

Change-Id: Ic60375f3e0ad7c7a8887dce7ffbd946f0a9d80b5
Reviewed-on: https://code.wireshark.org/review/209
Reviewed-by: Alexis La Goutte <alexis.lagoutte@gmail.com>
Tested-by: Alexis La Goutte <alexis.lagoutte@gmail.com>