Have git-compare-abis build in a sub-sub-subdirectory.
We have a lot of relative includes which apparently weren't updated
correctly when we moved the gtk directory to ui/gtk. For example,
ui/gtk/about_dlg.c has

The Autotools Makefile adds "-I../.." to the include path, which means
we end up including the current, top-level version_info.h instead of the
one in the compare-abis build directory.

Change-Id: I9df83857fc9f7d87a3b79d5b4da08d5e41625e4c
Reviewed-on: https://code.wireshark.org/review/436
Reviewed-by: Gerald Combs <gerald@wireshark.org>