Strip out carriage returns when reading version.conf.
The input record separator ($/) under Cygwin's Perl appears to be "\n"
which means that chomp() won't strip out carriage returns. Make sure
we do that when reading version.conf, otherwise we can end up with
unexpected behavior.

Check pkg_enable using an integer comparison as an extra level of
protection.

(We need to handle CRs better in general but for now we need to get
version stamping working for the next release.)

Change-Id: I19974467d1a7386728b69dca77f74e678d66ca6b
Reviewed-on: https://code.wireshark.org/review/483
Reviewed-by: Gerald Combs <gerald@wireshark.org>