Show ymsg.content as bytes; it's not a simple string.
The content of a YMSG message is a sequence of lines, each one of which
contains a text string (in some ASCII-based encoding) for a key, a
0xc080 separator, and a text string (in some ASCII-based encoding) for a
value.  That's not a string in any ASCII-based encoding I know of - 0xc0
0x80 is not, for example, a valid UTF-8 sequence (it's a too-long
sequence for NUL).

This should fix bug 9832:

	https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=9832

by avoiding the general "GTK+ on Windows crashes when asked to copy
something that's not valid UTF-8" problem.

Fix some field descriptions while we're at it.

Change-Id: I4084dabc89b0186ecd1a7329452ca2f1cb48f1c0
Reviewed-on: https://code.wireshark.org/review/488
Reviewed-by: Guy Harris <guy@alum.mit.edu>