Fix setting of STATS_LEN - it's what's in the file, not what we *generate*.
vwr->STATS_LEN is the length of the statistics trailer at the end of the
packet, and it's the same for S2 and S3 versions of the WLAN card.  It
should *not* be set to the length of the metadata headers that we
*generate and put in front of the packet data that we hand to our
callers*.

Get rid of a debugging message while we're at it.

Change-Id: I465b5ba4dedb88f1f401d34439b44b16a4bb01cc
Reviewed-on: https://code.wireshark.org/review/671
Reviewed-by: Guy Harris <guy@alum.mit.edu>