Fix bug9931 'Encapsulated ethernet packets sometimes show invalid FCS'
This fixes part-1 of bug9931: the uninitialized use of a wtap_pkthdr
struct. The second part of the bug deals with dissectors calling
the Ethernet dissector for ecnapsulated Ethernet packets but using
the wrong dissector handle to do so. That's unrelated to the issue this
commit addresses, so I'm splitting them up.

Cherry-Picked-From: I87be7b736f82dd74d8c261062f88143372b5344c
Reviewed-on: https://code.wireshark.org/review/848
Reviewed-by: Hadriel Kaplan <hadrielk@yahoo.com>
Reviewed-by: Anders Broman <a.broman58@gmail.com>

Conflicts:
	file.c
	frame_tvbuff.c
	reordercap.c
	ui/gtk/packet_list_store.c
	ui/qt/packet_list_model.cpp
	ui/tap_export_pdu.c

Change-Id: Ibc39d67e8a8717e70ddb453da3a868f6efb16bcc
Bug:10052
Reviewed-on: https://code.wireshark.org/review/2674
Reviewed-by: Evan Huus <eapache@gmail.com>